+++ R/pointSource.R
@@ -46,7 +46,7 @@ pointSource <- function(emissions, grid, verbose=T){
         xy <- xyFromCell(emis,id.cell)
         lon <- xy[1]
         lat <- xy[2]
-        print(paste("grid position","lat=",lat,"lon=",lon))
+        cat(paste("grid position","lat=",lat,"lon=",lon,"\n"))
       }
       emis[id.cell] <- emissions$e[i]
     }
@@ -69,22 +69,23 @@ pointSource <- function(emissions, grid, verbose=T){
         xy <- xyFromCell(emis,id.cell)
         lon <- xy[1]
         lat <- xy[2]
-        print(paste("Layer limits for grid position","lat=",lat,"lon=",lon))
+        cat(paste("Layer limits for grid position","lat=",lat,"lon=",lon,"\n"))
         for(k in 1:length(altura)){
-          print(paste("  height=", altura[k],"k=",k))
+          cat(paste("  height=",sprintf("%03.2f",altura[k]),"m k=",k,"\n"))
         }
       }
-      if(emissions$z[i] < altura[2]){
-        a     <- 1
-        print(paste("Emission heigh between", altura[1],"and",altura[2],"at k=",a,
-                    "for z=",emissions$z[i]))
-        cat("\n")
-      }
+      # if(emissions$z[i] < altura[2]){
+      #   a     <- 1
+      #   cat(paste("Emission heigh between", altura[1],"and",altura[2],"at k=",a,
+      #               "for z=",emissions$z[i]))
+      #   cat("\n")
+      # }
       for(k in 2:length(altura)){
         if(emissions$z[i] >= altura[k-1] & emissions$z[i] < altura[k]){
           a <- k -1
           if(verbose){
-            print(paste("Emission heigh between", altura[k-1],"and",altura[k],"at k=",a,
+            cat(paste("Emission heigh between", sprintf("%03.2f",altura[k-1]),"m and",
+                      sprintf("%03.2f",altura[k]),"m at k=",a,
                         "for z=",emissions$z[i]))
             cat("\n")
           }